byob/requirements.txt
mss==3.3.0
WMI==1.4.9
numpy==1.15.2
pyxhook==1.0.0
twilio==6.14.0
colorama==0.3.9
requests==2.20.0
paho-mqtt==1.5.0
PyInstaller==3.3.1
pycryptodomex==3.8.1
opencv-python==3.4.3.18;python_version<'3'
pyHook==1.5.1;sys.platform=='win32'
pypiwin32==223;sys.platform=='win32'
